+++ src/collectors/mod.rs
@@ -1,5 +1,6 @@
 pub mod bitcoin_core;
 mod error;
+pub(crate) mod helpers;
 pub mod lnd;
 pub mod registry;
 pub mod traits;